[Klib] Better fix for invalid coordinates
The compiler does not support reading klib with invalid coordinates
(such where start > end), so if those are ever encountered, the
serializer tries to "fix" them before producing a klib.

The current way of "fixing" encodes coordinates [10, 8] as [10, 12].
This produces "correct", but of course very different coordinates.
The new way of fixing encodes coordinates [10, 8] as [8, 10], which is
most likely the original intent.

There is also one case where coordinates with end < start might be
considered as valid, but still would be rejected by deserializer,
i.e. [N, -1] (-1 == INVALID_OFFSET). Those would also be "fixed" in the
old scheme to something completely different.
In the new scheme, we "fix" them to [-1, -1], which again, is quite
likely to capture the original intent.

At the same time we lift the restriction in deserializer about
end < start to possibly permit the [N, -1] case in the future.

KT-80910