K2/fixation: don't prioritize captured upper bounds in case type variable isn't bound to another unfixed variable
During the fix of KT-74999, we introduced a new readiness
READY_FOR_FIXATION_CAPTURED_UPPER_BOUND_WITH_SELF_TYPES.
that applied in case we have a constraint for non-reified variables
with self-based bounds and had higher priority than regular
LOWER/UPPER constraints. However, during investigation of KT-80577
it was concluded that such a fix is too broad and makes sense
only for a specific case with T <: Captured and T <: K. In this case
when K is fixed to some regular type T will be effectively
Captured & RegularType with high probability of being approximated
to Nothing. So in this commit we left only specific applications of
more prioritized READY_FOR_FIXATION_CAPTURED_UPPER_BOUND_WITH_SELF_TYPES.

#KT-80577 Fixed